requirements.txt
discord.py
pymongo
colour